NewAggregator2 -- move hardcoded domains/ii out of aggregator constructor
