+++ turbo/app/squeeze_cmd.go
@@ -137,7 +137,7 @@ func squeezeStorage(ctx context.Context, dirs datadir.Dirs, logger log.Logger) e
 	ac := agg.BeginFilesRo()
 	defer ac.Close()
 
-	aggOld, err := state.NewAggregator(ctx, dirsOld, config3.DefaultStepSize, db, logger)
+	aggOld, err := state.NewAggregator2(ctx, dirsOld, config3.DefaultStepSize, db, logger)
 	if err != nil {
 		panic(err)
 	}
@@ -178,7 +178,7 @@ func squeezeStorage(ctx context.Context, dirs datadir.Dirs, logger log.Logger) e
 func squeezeCode(ctx context.Context, dirs datadir.Dirs, logger log.Logger) error {
 	db := dbCfg(kv.ChainDB, dirs.Chaindata).MustOpen()
 	defer db.Close()
-	agg, err := state.NewAggregator(ctx, dirs, config3.DefaultStepSize, db, logger)
+	agg, err := state.NewAggregator2(ctx, dirs, config3.DefaultStepSize, db, logger)
 	if err != nil {
 		return err
 	}
